Changes to core files  allowing fitting using multiple protocols
These changes turns protocol into a cell of different protocols. variables used in fitting (experimental data, simulated current, ...) are replaced with a cell containing one for each protocol in the fitting procedure. The objective functions now loop over the protocols when calculating the value of the objective. When selecting parameter values in the objective function, r = CheckingRanges now operates on the r value from the maximum of the supplied protocols.