Added the CMAES and MCMC results from this project
Allows checking of the results. Note that the MCMC sampling does not seem to be working correctly in some cases as the chain looks to be nearly flat - I haven't managed to isolate why but believe its got something to do with the constraints on the rate constants, which remain essentially unchanged. In an mh model, these may need to be different.